Prevent re-rendering display options on search
